feat: pomodoro timer
Add a new command called `pomo` to manage pomodoro sessions based on the
Pomodoro Technique. The Pomodoro Technique is a simple yet effective
method for focused work with planned breaks in between. Francesco
Cirillo coined the term "pomodoro," which translates to "tomato," in the
late 1980s after the tomato-shaped timer he used as a university
student.